Moved web3 methods to separate modules
